requirements.txt
pyserial>=3.5
pynput>=1.7.0
loguru>=0.6.0
pystray>=0.19.0
pillow>=9.0.0
py2app>=0.28.0
PyQt5>=5.15.0
dmgbuild>=1.6.0
